refactor: move `are_we_wezterm()` function to mux.utils
